Add ContextTagResolver#ofUnsafe
